Update Car.java
I converted the Letter in VIN to ASCII, so it can go past Z to [ and beyond, but the code is simpler, and some other advancements in the constructer were done. My free period in my Latin class were half of the class is missing just ended so I need to wrap things up.